+++ lang/en/filter_synhi.php
@@ -23,7 +23,7 @@
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later.
  */
 
-$string['filtername'] = 'SynHi filter';
+$string['filtername'] = 'SynHi';
 
 // Settings.
 $string['engine'] = 'Engine';