+++ Changes.md
@@ -4,8 +4,9 @@ Version 3.7.0.4 - TBD
 1. Release candidiate.
 2. Add screenshots to Readme.md.
 3. Tidy code.
+4. Add information settings.
 
-Version 3.7.0.3 - 14/9/20
+Version 3.7.0.3 - 15/9/20
 1. Fix 'missing thirdpartylibs.xml' - #1.
 
 Version 3.7.0.2 - 3/9/20